# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Docs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/about/profile": {
            "get": {
                "tags": [
                    "About Me"
                ],
                "summary": "Get Profile",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "image": {
                                            "type": "string"
                                        },
                                        "first_name": {
                                            "type": "string"
                                        },
                                        "last_name": {
                                            "type": "string"
                                        },
                                        "sil_score": {
                                            "type": "integer"
                                        },
                                        "sil_score_progress": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        },
                                        "sessions_completed_count": {
                                            "type": "integer"
                                        },
                                        "events_scheduled_count": {
                                            "type": "integer"
                                        },
                                        "reflections_written_count": {
                                            "type": "integer"
                                        },
                                        "department": {
                                            "type": "string"
                                        },
                                        "role": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "phone": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "About Me"
                ],
                "summary": "Update profile. All properties is required.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User was successfully updated",
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/AboutMeUser"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/Validation Error"
                        }
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "About Me"
                ],
                "summary": "Update profile.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User was successfully updated",
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/AboutMeUser"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/Validation Error"
                        }
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            }
        },
        "/about/settings": {
            "get": {
                "tags": [
                    "About Me"
                ],
                "summary": "Get Settings",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/UserSettings"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "About Me"
                ],
                "summary": "Update Settings",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/UserSettings"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            }
        },
        "/about/share": {
            "patch": {
                "tags": [
                    "About Me"
                ],
                "summary": "Update ShareActivty Setting",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/UserSettings"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            }
        },
        "/about/photo": {
            "get": {
                "tags": [
                    "About Me"
                ],
                "summary": "Get Profile Photo",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "image": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "About Me"
                ],
                "summary": "Set Profile Photo",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "image": {
                                    "description": "Basse64 format",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "image": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "AboutMeUser": {
            "allOf": [
                {
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "first_name": {
                            "type": "string"
                        },
                        "last_name": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "client_id": {
                            "type": "integer"
                        },
                        "job_role": {
                            "type": "string"
                        },
                        "job_dept": {
                            "type": "string"
                        }
                    }
                },
                {
                    "$ref": "#/definitions/CreatedUpdatedDates"
                }
            ]
        },
...
        "CreatedUpdatedDates": {
            "properties": {
                "created_at": {
                    "description": "Format(PHP): Y:m:d H:i:s O",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Format(PHP): Y:m:d H:i:s O",
                    "type": "string"
                }
            }
        },
...
        "UserSettings": {
            "properties": {
                "tip_time": {
                    "type": "string"
                },
                "event_practice_email": {
                    "type": "integer"
                },
                "event_practice_notification": {
                    "type": "integer"
                },
                "random_acts_email": {
                    "type": "integer"
                },
                "random_acts_notification": {
                    "type": "integer"
                },
                "login_streak_email": {
                    "type": "integer"
                },
                "login_streak_notification": {
                    "type": "integer"
                },
                "share_activity": {
                    "type": "integer"
                }
            }
        },
...
        "UserBody": {
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "job_dept": {
                    "type": "string"
                },
                "job_role": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    },
...
    "securityDefinitions": {
        "apiToken": {
            "type": "apiKey",
            "description": "Api token",
            "name": "Authorization",
            "in": "header"
        }
    }
}